requirements.txt
pytorch
networkx
matplotlib
pylint
absl-py
skrl